fix replaceSupportEmail function to use environment setting
