+++ jsapp/js/components/languages/languagesListStore.ts
@@ -3,7 +3,7 @@ import type {
   PaginatedResponse,
   FailResponse,
 } from 'js/dataInterface';
-import {handleApiFail} from 'js/utils';
+import {handleApiFail} from 'js/api';
 import {ROOT_URL} from 'js/constants';
 import languagesStore from './languagesStore';
 import type {ListLanguage} from './languagesStore';
