+++ jsapp/js/utils.ts
@@ -13,8 +13,8 @@ import {toast} from 'react-hot-toast';
 import {Cookies} from 'react-cookie';
 // importing whole constants, as we override ROOT_URL in tests
 import constants from 'js/constants';
-import type {FailResponse} from './dataInterface';
 import type Raven from 'raven';
+import envStore from 'js/envStore';
 
 export const LANGUAGE_COOKIE_NAME = 'django_language';
 
@@ -81,51 +81,6 @@ notify.warning = (msg: Toast['message'], opts?: ToastOptions): Toast['id'] =>
   notify(msg, 'warning', opts);
 notify.success = (msg: Toast['message'], opts?: ToastOptions): Toast['id'] =>
   notify(msg, 'success', opts);
-
-/**
- * Useful for handling the fail responses from API. Its main goal is to display
- * a helpful error toast notification and to pass the error message to Raven.
- *
- * It can detect if we got HTML string as response and uses a generic message
- * instead of spitting it out.
- */
-export function handleApiFail(response: FailResponse) {
-  // Avoid displaying toast when purposefuly aborted a request
-  if (response.status === 0 && response.statusText === 'abort') {
-    return;
-  }
-
-  let message = response.responseText;
-
-  // Detect if response is HTML code string
-  if (
-    typeof message === 'string' &&
-    message.includes('</html>') &&
-    message.includes('</body>')
-  ) {
-    // Try plucking the useful error message from the HTML string - this works
-    // for Werkzeug Debugger only. It is being used on development environment,
-    // on production this would most probably result in undefined message (and
-    // thus falling back to the generic message below).
-    const htmlDoc = new DOMParser().parseFromString(message, 'text/html');
-    message = htmlDoc.getElementsByClassName('errormsg')?.[0]?.innerHTML;
-  }
-
-  if (!message) {
-    message = t('An error occurred');
-    if (response.status || response.statusText) {
-      message += ` — ${response.status} ${response.statusText}`;
-    } else if (!window.navigator.onLine) {
-      // another general case — the original fetch response.message might have
-      // something more useful to say.
-      message += ' — ' + t('Your connection is offline');
-    }
-  }
-
-  notify.error(message);
-
-  window.Raven?.captureMessage(message);
-}
 
 /**
  * Returns a copy of arr with separator inserted in every other place.
@@ -262,9 +217,13 @@ const originalSupportEmail = 'help@kobotoolbox.org';
 // TODO: make this use environment endpoint's `support_email` property.
 // Currently no place is using this correctly.
 // See: https://github.com/kobotoolbox/kpi/issues/3910
-export function replaceSupportEmail(str: string, newEmail?: string): string {
-  if (typeof newEmail === 'string') {
-    return str.replace(originalSupportEmail, newEmail);
+/**
+ * Replaces the hardcoded email string (coming from transifex translation) with
+ * the one from the `/environment` endpoint.
+ */
+export function replaceSupportEmail(str: string): string {
+  if (typeof envStore.data.support_email === 'string' && envStore.data.support_email.length !== 0) {
+    return str.replace(originalSupportEmail, envStore.data.support_email);
   } else {
     return str;
   }
